exercicio 22

+++ ex10.py
@@ -1,3 +1,3 @@
 dinheiro = float(input('Digite quanto você tem de dinheiro'))
 
-print('Voce tem R${} e convertendo tem uss{}'.format(dinheiro, int(dinheiro/3,25)))
+print('Voce tem R${} e convertendo tem uss{}'.format(dinheiro, dinheiro/3,25))
